Comment and cleanup
